+++ src/OpenTelemetry.Exporter.InMemory/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Changed error handling, `InMemoryExporter` will now throw
+  `ObjectDisposedException` if `Export` is invoked after the exporter is
+  disposed.
+
 ## 1.4.0-alpha.2
 
 Released 2022-Aug-18